+++ src/test/java/BaseTest.java
@@ -12,57 +12,20 @@ import org.testng.annotations.*;
 
 import java.time.Duration;
 
-
 public class BaseTest {
     public WebDriver driver;
     WebDriverWait wait;
-
-
+    String url = "https://bbb.testpro.io";
 
     @BeforeMethod
     public void setUpBrowser() {
         WebDriverManager.chromedriver().setup();
         driver = new ChromeDriver();
-        wait = new WebDriverWait(driver, Duration.ofSeconds(4));
-        String url = "https://bbb.testpro.io";
         driver.get(url);
         driver.manage().window().maximize();
     }
-
     @AfterMethod
     public void tearDown() {
-
-        //driver.quit();
+        driver.quit();
     }
-
-
-
-//delete after
-
-    public void openLoginUrl() {
-        String url = "https://bbb.testpro.io";
-        driver.get(url);
-    }
-    public void enterEmail(String email) {
-        WebElement emailfield = driver.findElement(By.cssSelector("input[type='email']"));
-        emailfield.sendKeys(email);
-    }
-
-    public void enterPassword(String password) {
-        WebElement passwordfield = driver.findElement(By.cssSelector("input[type='password']"));
-        passwordfield.sendKeys(password);
-    }
-
-    public void submit() {
-        WebElement submitbtn = driver.findElement(By.cssSelector("button[type='submit']"));
-        submitbtn.click();
-    }
-
-    public void login() {
-        enterEmail("maxim.ibon@gmail.com");
-        enterPassword("te$t$tudent1");
-        submit();
-    }
-
-
 }
