+++ src/test/java/Pages/AllSongsPage.java
@@ -1,0 +1,9 @@
+package Pages;
+
+import org.openqa.selenium.WebDriver;
+
+public class AllSongsPage extends BasePage{
+    public AllSongsPage(WebDriver givenDriver) {
+        super(givenDriver);
+    }
+}
