+++ src/test/java/Pages/HomePage.java
@@ -1,0 +1,9 @@
+package Pages;
+
+import org.openqa.selenium.WebDriver;
+
+public class HomePage extends BasePage{
+    public HomePage(WebDriver givenDriver) {
+        super(givenDriver);
+    }
+}
