+++ src/test/resources/TestNG.xml
@@ -6,15 +6,10 @@
         <classes>
             <class name="LoginTests"/>
             <class name="BaseTest"/>
-            <class name="Homework16"/>
+           <class name="Homework16"/>
             <class name="Homework17"/>
             <class name="ProfileTest"/>
-            <class name="Homework18">
-            <methods>
-                <include name="playSong"/>
-                <include name="clickPlayBtn" />
-            </methods>
-            </class>
+            <class name="Homework18"/>
             <class name="Homework19"/>
             <class name="Homework20"/>
             <class name="Homework21"/>
